# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TrekMarket Nepal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /treks/{trekId}/join-requests:
    get:
      tags: [joinRequests]
      operationId: listJoinRequests
      summary: List join requests for a trek (agency only)
      parameters:
        - name: trekId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Join requests list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JoinRequest'
    post:
      tags: [joinRequests]
      operationId: createJoinRequest
      summary: Request to join a trek (trekker only)
      parameters:
        - name: trekId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJoinRequestBody'
      responses:
        '201':
          description: Join request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinRequest'
        '401':
          description: Unauthorized
  /join-requests/{requestId}:
    get:
      tags: [joinRequests]
      operationId: getJoinRequest
      summary: Get a single join request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Join request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinRequest'
    put:
      tags: [joinRequests]
      operationId: updateJoinRequest
      summary: Accept or reject a join request (agency only)
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJoinRequestBody'
      responses:
        '200':
          description: Updated join request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinRequest'
        '403':
          description: Forbidden
  /custom-requests:
    get:
      tags: [customRequests]
      operationId: listCustomRequests
      summary: List custom trek requests (agencies see all open; trekkers see their own)
      responses:
        '200':
          description: Custom requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomRequest'
    post:
      tags: [customRequests]
      operationId: createCustomRequest
      summary: Create a custom trek request (trekker only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomRequestBody'
      responses:
        '201':
          description: Created custom request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomRequest'
        '401':
          description: Unauthorized
  /custom-requests/{requestId}:
    get:
      tags: [customRequests]
      operationId: getCustomRequest
      summary: Get custom request details
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Custom request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomRequest'
        '404':
          description: Not found
  /custom-requests/{requestId}/bids:
    get:
      tags: [bids]
      operationId: listBids
      summary: List bids on a custom request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bids list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bid'
    post:
      tags: [bids]
      operationId: createBid
      summary: Submit a bid on a custom request (agency only)
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBidBody'
      responses:
        '201':
          description: Bid created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bid'
        '401':
          description: Unauthorized
  /bids/{bidId}/select:
    post:
      tags: [bids]
      operationId: selectBid
      summary: Select a bid (trekker only — closes the custom request)
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bid selected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bid'
        '403':
          description: Forbidden
  /bookings:
    get:
      tags: [bookings]
      operationId: listMyBookings
      summary: List current user's bookings
      responses:
        '200':
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
    post:
      tags: [bookings]
      operationId: createBooking
      summary: Create a booking with mock payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingBody'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '401':
          description: Unauthorized
  /bookings/{bookingId}:
    get:
      tags: [bookings]
      operationId: getBooking
      summary: Get booking details
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Not found
    delete:
      tags: [bookings]
      operationId: cancelBooking
      summary: Cancel a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
...
components:
...
  schemas:
...
    Trek:
      type: object
      required: [id, agencyId, title, destination, duration, startDate, price, maxGroupSize, description, status, currentParticipants, difficultyLevel, createdAt]
      properties:
        id:
          type: string
        agencyId:
          type: string
        agencyName:
          type: [string, 'null']
        title:
          type: string
        destination:
          type: string
        duration:
          type: integer
          description: Duration in days
        startDate:
          type: string
          format: date
        endDate:
          type: [string, 'null']
          format: date
        price:
          type: number
        maxGroupSize:
          type: integer
        description:
          type: string
        imageUrl:
          type: [string, 'null']
        status:
          type: string
          enum: [active, cancelled, completed]
        currentParticipants:
          type: integer
        difficultyLevel:
          type: string
          enum: [easy, moderate, hard, extreme]
        createdAt:
          type: string
          format: date-time
...
    JoinRequest:
      type: object
      required: [id, trekId, trekkerId, status, createdAt]
      properties:
        id:
          type: string
        trekId:
          type: string
        trekkerId:
          type: string
        trekkerName:
          type: [string, 'null']
        trekkerEmail:
          type: [string, 'null']
        trekkerProfileImage:
          type: [string, 'null']
        trek:
          $ref: '#/components/schemas/Trek'
        status:
          type: string
          enum: [pending, accepted, rejected]
        message:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    CreateJoinRequestBody:
      type: object
      properties:
        message:
          type: string
    UpdateJoinRequestBody:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected]
    CustomRequest:
      type: object
      required: [id, trekkerId, destination, budget, startDate, groupSize, status, createdAt]
      properties:
        id:
          type: string
        trekkerId:
          type: string
        trekkerName:
          type: [string, 'null']
        destination:
          type: string
        budget:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: [string, 'null']
          format: date
        groupSize:
          type: integer
        notes:
          type: [string, 'null']
        status:
          type: string
          enum: [open, closed, cancelled]
        bidsCount:
          type: integer
        selectedBidId:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    CreateCustomRequestBody:
      type: object
      required: [destination, budget, startDate, groupSize]
      properties:
        destination:
          type: string
        budget:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        groupSize:
          type: integer
        notes:
          type: string
    Bid:
      type: object
      required: [id, customRequestId, agencyId, proposedPrice, planDescription, status, createdAt]
      properties:
        id:
          type: string
        customRequestId:
          type: string
        agencyId:
          type: string
        agencyName:
          type: [string, 'null']
        agencyProfileImage:
          type: [string, 'null']
        proposedPrice:
          type: number
        planDescription:
          type: string
        message:
          type: [string, 'null']
        status:
          type: string
          enum: [pending, selected, rejected]
        createdAt:
          type: string
          format: date-time
    CreateBidBody:
      type: object
      required: [proposedPrice, planDescription]
      properties:
        proposedPrice:
          type: number
        planDescription:
          type: string
        message:
          type: string
    Booking:
      type: object
      required: [id, trekkerId, totalAmount, advanceAmount, status, createdAt]
      properties:
        id:
          type: string
        trekkerId:
          type: string
        trekId:
          type: [string, 'null']
        bidId:
          type: [string, 'null']
        trek:
          $ref: '#/components/schemas/Trek'
        totalAmount:
          type: number
        advanceAmount:
          type: number
        status:
          type: string
          enum: [pending, paid, cancelled]
        paymentRef:
          type: [string, 'null']
        cancellationPolicy:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    CreateBookingBody:
      type: object
      properties:
        trekId:
          type: string
        bidId:
          type: string
...