# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrekMarket Nepal API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";
import type { Trek } from "./trek";

export interface Booking {
  id: string;
  trekkerId: string;
  /** @nullable */
  trekId?: string | null;
  /** @nullable */
  bidId?: string | null;
  trek?: Trek;
  totalAmount: number;
  advanceAmount: number;
  status: BookingStatus;
  /** @nullable */
  paymentRef?: string | null;
  /** @nullable */
  cancellationPolicy?: string | null;
  createdAt: Date;
}