# lib/api-zod/src/generated/types/trek.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrekMarket Nepal API
 * OpenAPI spec version: 0.1.0
 */
import type { TrekDifficultyLevel } from "./trekDifficultyLevel";
import type { TrekStatus } from "./trekStatus";

export interface Trek {
  id: string;
  agencyId: string;
  /** @nullable */
  agencyName?: string | null;
  title: string;
  destination: string;
  /** Duration in days */
  duration: number;
  startDate: Date;
  /** @nullable */
  endDate?: Date | null;
  price: number;
  maxGroupSize: number;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  status: TrekStatus;
  currentParticipants: number;
  difficultyLevel: TrekDifficultyLevel;
  createdAt: Date;
}